+++ CHANGELOG.md
@@ -16,6 +16,7 @@ This project adheres to [Semantic Versioning](http://semver.org/) and [Keep a Ch
 - [#1799](https://gitlab.com/meltano/meltano/issues/1799) Update "Date Range(s)" button label to account for pluralization
 - [#1799](https://gitlab.com/meltano/meltano/issues/1799) Fallback to inline text and only display the date range `<select>` if there are two or more date ranges to filter on
 - [#1799](https://gitlab.com/meltano/meltano/issues/1799) Update date range picker to initialize at the first attribute with a valid date range
+- [#1799](https://gitlab.com/meltano/meltano/issues/1799) Update the Report Builder's "jump to date range dropdown" buttons (small calendar icon button associated with the left pane's attribute items) to automatically focus the date range that's associated
 
 ### Fixes